requirements.txt
accelerate==0.28.0
decord==0.6.0
deepspeed==0.12.6
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@afe9eb920646102f7e6bf0cd2115841cea2aca13
diffusers==0.26.3
einops==0.7.0
fastapi==0.110.1
ffmpy==0.3.2
flash-attn==2.5.6
ftfy==6.2.0
future==1.0.0
gradio==3.42.0
gradio_client==0.5.0
huggingface-hub==0.22.2
hydra-core==1.3.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.9
Jinja2 @ file:///croot/jinja2_1706733616596/work
kornia==0.7.2
kornia_rs==0.1.2
matplotlib==3.8.3
ninja==1.11.1.1
nltk==3.8.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
omegaconf==2.3.0
open-clip-torch==2.23.0
openai==0.27.8
opencv-python==4.9.0.80
pandas==2.2.1
peft==0.10.0
pillow @ file:///croot/pillow_1707233021655/work
pycocotools==2.0.7
pydantic==2.6.4
pydantic_core==2.16.3
pytorch-lightning==1.5.0
pytorchvideo==0.1.5
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2023.12.25
requests @ file:///croot/requests_1707355572290/work
rpds-py==0.18.0
safetensors==0.4.2
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==0.9.16
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.1
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.3.2
torchvision @ file:///workspace/croot/torchvision_1693425165028/work
tqdm==4.66.2
transformers==4.31.0
websockets==11.0.3
Werkzeug==3.0.2
whisper==1.1.10
